requirements-coordination.txt
# Requirements for Docker Coordination System
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
pika==1.3.2
aiohttp==3.8.6
aioredis==2.0.1
websockets==11.0.3
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Web3 and blockchain
web3==6.11.3
eth-account==0.9.0
eth-utils==2.3.0

# Docker SDK
docker==6.1.3

# LangChain - Compatible versions that resolve dependency conflicts
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-experimental==0.0.68

# AutoGen (optional) - Compatible with openai>=1.0
autogen-agentchat==0.4.0

# AI/ML dependencies - Use openai>=1.0 for compatibility with newer langchain and autogen
openai==1.55.0
anthropic==0.7.7
ollama==0.1.7

# Utilities
requests==2.31.0
pydantic==2.5.0
typing-extensions==4.8.0
python-json-logger==2.0.7
structlog==23.2.0

# Monitoring and observability
prometheus-client==0.19.0
grafana-api==1.0.3

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Security
cryptography==41.0.7
pyjwt==2.8.0
